+++ src/main/java/com/github/nthykier/debpkg/dch/DchElementFactory.java
@@ -1,9 +1,6 @@
 package com.github.nthykier.debpkg.dch;
 
-import com.github.nthykier.debpkg.dch.psi.DchChangelogEntry;
-import com.github.nthykier.debpkg.dch.psi.DchFile;
-import com.github.nthykier.debpkg.dch.psi.DchSignoff;
-import com.github.nthykier.debpkg.dch.psi.DchSignoffDate;
+import com.github.nthykier.debpkg.dch.psi.*;
 import com.github.nthykier.debpkg.deb822.psi.Deb822FieldValuePair;
 import com.github.nthykier.debpkg.deb822.psi.Deb822Substvar;
 import com.github.nthykier.debpkg.deb822.psi.Deb822ValueParts;
@@ -22,9 +19,19 @@ public class DchElementFactory {
         return (DchFile)PsiFileFactory.getInstance(project).createFileFromText(name, DchFileType.INSTANCE, text);
     }
 
+    public static DchChangelogEntry createChangelogEntry(Project project, String text) {
+        return Objects.requireNonNull((DchChangelogEntry)createFile(project, text).getFirstChild());
+    }
+
     public static DchSignoff createSignoff(Project project, String text) {
-        DchChangelogEntry changelogEntry = (DchChangelogEntry)createFile(project, text).getFirstChild();
+        DchChangelogEntry changelogEntry = createChangelogEntry(project, text);
         return changelogEntry.getSignoff();
+    }
+
+    @NotNull
+    public static DchSource createSource(Project project, String source) {
+        String text = source + " (1.0) unstable; urgency=low\n * foo\n -- Someone <somewhare@example.com>  Sat, 25 Sep 2021 06:58:09 +0000";
+        return Objects.requireNonNull(createChangelogEntry(project, text).getVersionLine().getSource());
     }
 
     @NotNull